refactor: relocate typing event logic to the websocket service and add linkifyjs dependency.
